% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdg_norm_lmx.R
\name{bdg_norm_lmx}
\alias{bdg_norm_lmx}
\title{bdg_norm_lmx (Reference sample based normalization)}
\usage{
bdg_norm_lmx(
  bridge.ls,
  data.ls,
  between.plate.method = "max",
  round_digits = 3,
  from_assay = "default",
  save_assay = "normed"
)
}
\arguments{
\item{bridge.ls}{list of bridge.summarizedexperiment objs using pull_bdg()}

\item{data.ls}{list of summarizedexperiment objs using read_npx().
normalized assay slot will be added to each of the object in the list.}

\item{between.plate.method}{method to set the inter plate reference value of bridging samples
using max can garantee no negative value of the normalized data}

\item{round_digits}{digit kept}

\item{from_assay}{select assay slot to be normalized}

\item{save_assay}{name the assay slot for normalized data}
}
\description{
normalize by adjust samples' value per plate to make that
the reference sample among all plates are the same. option to
pick using median, max, mean
}
