% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdg_norm_lmx_multi.R
\name{bdg_norm_lmx_multi}
\alias{bdg_norm_lmx_multi}
\title{bdg_norm_lmx_multi (Reference sample based normalization)}
\usage{
bdg_norm_lmx_multi(
  bridge.str,
  data.ls,
  round_digits = 3,
  between.plate.method = "mean",
  from.assay = "default",
  save_assay = "normed"
)
}
\arguments{
\item{bridge.str}{vector of common strings to identify sets of bridging samples}

\item{data.ls}{list of summarizedexperiment objs using read_npx().
normalized assay slot will be added to each of the object in the list.}

\item{round_digits}{digit kept}

\item{save_assay}{name the assay slot for normalized data}

\item{from_assay}{select assay slot to be normalized}
}
\description{
normalize by adjust samples' value per plate to make that
the reference sample among all plates are the same. option to
pick using median, max, mean
}
