% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmb_lmx_se.R
\name{cmb_lmx_se}
\alias{cmb_lmx_se}
\title{cmb_lmx_se}
\usage{
cmb_lmx_se(se_list)
}
\arguments{
\item{se_list}{list of npx summarizedexperiment obj
rowData will be cbind, common colData}
}
\value{
combined summarizedexperiment obj
}
\description{
combind a list of luminex summarizedexperiment
}
