% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_qc_plot.R
\name{count_qc_plot}
\alias{count_qc_plot}
\title{count_qc_plot}
\usage{
count_qc_plot(f_list, pre_fix)
}
\arguments{
\item{pre_fix}{pre_fix for the csv to be saved.}

\item{x}{component object created by read_lmx_csv()}
}
\description{
Heatmap plot count location to visualize the
location with beads count < 35. red color indicate
low beads count
}
