% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmx_norm_from_raw.R
\name{lmx_norm_from_raw}
\alias{lmx_norm_from_raw}
\title{lmx_norm_from_raw}
\usage{
lmx_norm_from_raw(
  f_list,
  plex,
  pre_fix,
  map_csv,
  morm_method = "median",
  ref_sample = "QC"
)
}
\arguments{
\item{plex}{"384" or "96"}

\item{f}{a string file path}
}
\value{
a list each elemment is a component from the raw csv.
}
\description{
read lmx raw file to list of dataframes of each components
}
