% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_replace_fun.R
\name{norm_replace_fun}
\alias{norm_replace_fun}
\title{norm_replace_fun}
\usage{
norm_replace_fun(f_list, morm_method = "median", ref_sample = "QC", map_csv)
}
\arguments{
\item{f_list}{a named list created from read_lmx_csv()}

\item{morm_method}{normalization method, such as mean, median...}

\item{ref_sample}{reference sample for between plates normalization}

\item{map_csv}{csv files stored in inst/extra, or user}
}
\value{
a list with the MFI component ref-sample normalized.
}
\description{
normalized MFI component by selected reference sample,
position-wise average the STD from each plates,
only the 1st plate will keep STD as the averaged value,
(this is adapt to the luminex analyser's format).
}
