% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lmx_qc.R
\name{plot_lmx_qc}
\alias{plot_lmx_qc}
\title{plot_lmx_qc (Generate QC plots pre and post normalization)}
\usage{
plot_lmx_qc(normed_se, bridge_pattern = "HD Urine Pool", fields = "Sample")
}
\arguments{
\item{normed_se}{a summarizedexperiment, normaly generated by cmb_npx_sc() after normalization}

\item{bridge_pattern}{regex to identify bridging samples}

\item{fields}{colData name from which to identify bridging samples uing the bridge_pattern}
}
\description{
normalization QC plots, generate
1) analyt/pre vs post scatter plot, colored by plates
2) optional to select a analyt of choice plot for
pre vs post npx by plates scatter plots
3) optional plot to track Randox behavior pre and post normalization
if randox_info is given plot the targeted value
}
