% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lmx.R
\name{read_lmx}
\alias{read_lmx}
\title{read_lmx}
\usage{
read_lmx(f, lot = "default", analyt_start_row)
}
\arguments{
\item{f}{a string file path}

\item{analyt_start_row}{8 for npx, 9 for quant}

\item{lot:}{optional parameter to record lot information}
}
\value{
a summarizedexperimet object.
rowData is the feature and HOD, and LOD,
colData is the unique_id and subject id given by the operator, and
assay file name where the sample was assayed
}
\description{
read lmx raw file to create a summarizedexperiment
>HOD is replaced by Inf, <LOD is replaced by -Inf;
retrieve STD3, and the lowest, the hightest STD's MFI, Conc.
(correlated to LLOD and HLOD and matching MFI)
}
