% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lmx_csv.R
\name{read_lmx_csv}
\alias{read_lmx_csv}
\title{read_lmx_csv}
\usage{
read_lmx_csv(f, plex)
}
\arguments{
\item{f}{a string file path}

\item{plex}{"384" or "96"}
}
\value{
a list each elemment is a component from the raw csv.
}
\description{
read lmx raw file to list of dataframes of each components
}
