% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_comp2csv.R
\name{save_comp2csv}
\alias{save_comp2csv}
\title{save_comp2csv}
\usage{
save_comp2csv(f_list, pre_fix = "~/Downloads/")
}
\arguments{
\item{f_list}{a named list created from read_lmx_csv()}

\item{pre_fix}{pre_fix for the csv to be saved.}
}
\description{
save normalized component list to csv
}
