% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_dilute.R
\name{update_dilute}
\alias{update_dilute}
\title{update samples with different dilution factors}
\usage{
update_dilute(se, sample_pattern = "QC", fields = "Sample", dilute.f = 1)
}
\arguments{
\item{se}{npx summarizedexpermient object}

\item{sample_pattern}{string pattern to sample}

\item{fields}{colname in metadata to filter sample}

\item{dilute.f}{numeric dilution factor}
}
\description{
update samples with different dilution factors
}
